@@ -15,18 +15,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.0.0",
+    "@angular/animations": "12.1.1",
     "@angular/cdk": "^8.0.0",
-    "@angular/common": "8.0.0",
-    "@angular/compiler": "8.0.0",
-    "@angular/core": "8.0.0",
+    "@angular/common": "12.1.1",
+    "@angular/compiler": "12.1.1",
+    "@angular/core": "12.1.1",
     "@angular/fire": "^5.2.1",
-    "@angular/forms": "8.0.0",
+    "@angular/forms": "12.1.1",
     "@angular/material": "^8.0.0",
-    "@angular/platform-browser": "8.0.0",
-    "@angular/platform-browser-dynamic": "8.0.0",
-    "@angular/router": "8.0.0",
-    "@angular/service-worker": "8.0.0",
+    "@angular/platform-browser": "12.1.1",
+    "@angular/platform-browser-dynamic": "12.1.1",
+    "@angular/router": "12.1.1",
+    "@angular/service-worker": "12.1.1",
     "@ngx-translate/core": "11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "core-js": "^2.6.9",
@@ -34,15 +34,15 @@
     "firebase": "^5.11.1",
     "hammerjs": "^2.0.8",
     "ng-lazyload-image": "^5.1.2",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.6.7",
     "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.800.0",
     "@angular/cli": "8.0.1",
-    "@angular/compiler-cli": "8.0.0",
-    "@angular/language-service": "8.0.0",
+    "@angular/compiler-cli": "12.1.1",
+    "@angular/language-service": "12.1.1",
     "@types/jasmine": "2.8.9",
     "@types/node": "~8.0.30",
     "codelyzer": "^5.0.1",
@@ -59,7 +59,7 @@
     "protractor": "~5.4.2",
     "ts-node": "~5.0.1",
     "tslint": "~5.11.0",
-    "typescript": "3.4.5",
+    "typescript": "4.3.5",
     "webpack-bundle-analyzer": "^3.3.2"
   }
 }
