@@ -36,7 +36,7 @@ import { PaginationComponent } from './component/pagination/pagination.component
   ],
 })
 export class SharedModule {
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<SharedModule> {
     return {
       ngModule: SharedModule,
       providers: [
